% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{add_cell_ids}
\alias{add_cell_ids}
\title{Add cell IDs to a h5_list object}
\usage{
add_cell_ids(
  h5_list,
  target = "matrix",
  add_uuid = TRUE,
  replace_barcode = TRUE,
  retain_original_barcode = TRUE,
  add_name = TRUE
)
}
\arguments{
\item{h5_list}{a list object generated by running rhdf5::h5dump() on a 10x HDF5 file.}

\item{target}{the group name for the matrix within the h5_list to use. Default is "matrix", which is used by 10x.}

\item{add_uuid}{a logical indicating whether or not to add a uuid generated by ids::uuid(). Default is TRUE.}

\item{replace_barcode}{a logical indicating whether or not the new uuid should replace the barcodes slot. Default is TRUE.}

\item{retain_original_barcode}{a logical indicating whether to retain the original barcodes if replace_barcode is TRUE. These will be moved to /matrix/original_barcode if TRUE. Default is TRUE.}

\item{add_name}{a logical indicating whether to add a new cell name. These are adjective_adjective_animal names generated by ids::adjective_animal(). Default is TRUE.}
}
\value{
a modified h5_list object
}
\description{
Add cell IDs to a h5_list object
}
