% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{add_h5_list_hash_results}
\alias{add_h5_list_hash_results}
\title{Add cell hashing results to an h5_list.}
\usage{
add_h5_list_hash_results(
  h5_list,
  hash_category_table,
  hash_count_matrix,
  match_target = "barcodes"
)
}
\arguments{
\item{h5_list}{a list object generated by running rhdf5::h5dump() on a 10x HDF5 file.}

\item{hash_category_table}{a data.frame or data.table with results from HTOparser::categorize_binary_hash_matrix() consisting of 3 columns: cell_barcode, hto_category, and hto_barcode.}

\item{hash_count_matrix}{(optional) a matrix or dgCMatrix with results from CITE-seq-Count, HTOparser::add_missing_hto_rows(), or HTOparser::barcode_table_to_matrix().}

\item{match_target}{a character object specifying which values to use for filtering in the h5_list. If "barcodes", uses matrix/barcodes; Otherwise, looks in matrix/observations for matching values}
}
\value{
a list object with hash data added to /hash/ and /matrix/observations/
}
\description{
Add cell hashing results to an h5_list.
}
