% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rna_h5.R
\name{add_h5_transposed_matrix}
\alias{add_h5_transposed_matrix}
\title{Add a transposed version of a matrix to an .h5 file}
\usage{
add_h5_transposed_matrix(
  h5_file,
  in_target = "matrix",
  in_feature_names = "id",
  in_sample_names = "barcodes",
  out_target = paste0("transposed_", target)
)
}
\arguments{
\item{h5_file}{An existing .h5 file.}

\item{in_target}{The name of the target matrix. Default is "matrix".}

\item{in_feature_names}{The feature names to use in the transposed matrix. Default is "id".}

\item{in_sample_names}{The observation names to use in the transposed matrix. Default is "barcodes".}

\item{out_target}{The name of the output matrix to store. Default is "transposed_" prefixed to the in_target param (e.g. "transposed_matrix").}
}
\value{
No return
}
\description{
Useful for applications that need feature-indexed, rather than observation-indexed values.
}
