% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_matrix_analysis.R
\name{add_missing_hto_rows}
\alias{add_missing_hto_rows}
\title{Add zero values for missing HTOs to a matrix}
\usage{
add_missing_hto_rows(mat, valid_htos)
}
\arguments{
\item{mat}{A matrix or sparse matrix of hash tag oligo counts. Rows must be hashes, columns must be cells}

\item{valid_htos}{A character vector of valid hto sequences.}
}
\value{
a dgCMatrix of HTO counts with rows equal to valid_htos.
}
\description{
Add zero values for missing HTOs to a matrix
}
