% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{add_well_metadata}
\alias{add_well_metadata}
\title{Add well metadata}
\usage{
add_well_metadata(h5_list, well_id)
}
\arguments{
\item{h5_list}{a list object generated by running rhdf5::h5dump() on a 10x HDF5 file.}

\item{well_id}{a character object indicating the well name for the input h5. Should match the pattern B000-P1C1W1}
}
\value{
a modified h5_list object
}
\description{
Parses a standard well_id to generate chip_id, pool_id, and batch_id.
}
