% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rna_h5.R
\name{append_ext_h5_character}
\alias{append_ext_h5_character}
\title{Append character values to an existing HDF5 string dataset}
\usage{
append_ext_h5_character(object, h5_handle, name)
}
\arguments{
\item{object}{a character vector to write.}

\item{h5_handle}{An HDF5 connection handle generated by rhdf5::H5Fopen()}

\item{name}{The name of the location to append to within the HDF5 file.}
}
\value{

}
\description{
Append character values to an existing HDF5 string dataset
}
