% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_matrix_analysis.R
\name{binarize_hash}
\alias{binarize_hash}
\title{Binarize a vector of hash counts based on a cutoff value}
\usage{
binarize_hash(x, cutoff)
}
\arguments{
\item{x}{a numeric vector of hash count values}

\item{cutoff}{a numeric value to use as a cutoff}
}
\value{
a binary vector with value 1 for all values in x > cutoff
}
\description{
Binarize a vector of hash counts based on a cutoff value
}
