% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{cat_h5_list}
\alias{cat_h5_list}
\title{Concatenate two h5_list objects.}
\usage{
cat_h5_list(x, y)
}
\arguments{
\item{x}{the first list to combine.}

\item{y}{the second list to append.}
}
\value{
a list object with the same structure as x.
}
\description{
This is suitable for h5_list objects that contain vectors and/or dgCMatrices. It expects that the two lists have the same structure and names.
}
