% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_matrix_analysis.R
\name{categorize_binary_hash_matrix}
\alias{categorize_binary_hash_matrix}
\title{Convert binary hash matrix results to categorical labels for each cell barcode}
\usage{
categorize_binary_hash_matrix(bmat)
}
\arguments{
\item{bmat}{a matrix or sparse matrix with binarized hash scores, as generated by `binarize_hash_matrix()`. Rows must be hash barcodes, and columns must be cell barcodes.}
}
\value{
A list containing two objects:
\itemize{
  \item hash_category_table: a data.frame of categorical hash results: cell_barcode, hash_category, hash_sequence
  \item hash_summary: a data.frame with summary statistics for hashes in this run: hash_category, n_category, frac_category
}
}
\description{
Convert binary hash matrix results to categorical labels for each cell barcode
}
