% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_matrix}
\alias{check_matrix}
\title{Simple function to check for matrix or dgCMatrix classes for assertions}
\usage{
check_matrix(x)
}
\arguments{
\item{x}{an object to check for matrix or dgCMatrix classes}
}
\value{
a logical value
}
\description{
Simple function to check for matrix or dgCMatrix classes for assertions
}
