% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rna_h5.R
\name{choose_integer_bits}
\alias{choose_integer_bits}
\title{Decide how many bits to use to store integer values in an HDF5 Dataset object.}
\usage{
choose_integer_bits(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
an integer value. One of 16L, 32L, 64L, or NULL if larger than 64-bit.
}
\description{
Decide how many bits to use to store integer values in an HDF5 Dataset object.
}
