% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{convert_char_na}
\alias{convert_char_na}
\title{Convert "NA" character entries to actual NAs}
\usage{
convert_char_na(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character vector with NAs
}
\description{
Convert "NA" character entries to actual NAs
}
