% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{convert_char_na_recursive}
\alias{convert_char_na_recursive}
\title{Convert all "NA" character values to actual NAs recursively}
\usage{
convert_char_na_recursive(x)
}
\arguments{
\item{x}{a list object}
}
\value{
a list object with "NA"s converted to NA
}
\description{
Convert all "NA" character values to actual NAs recursively
}
