% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rna_h5.R
\name{create_ext_h5_character}
\alias{create_ext_h5_character}
\title{Create an extensible character HDF5 dataset}
\usage{
create_ext_h5_character(h5_handle, name, size = 32)
}
\arguments{
\item{h5_handle}{a connection object created by H5Fopen()}

\item{name}{a character object specifying the full path to the object to modify within the HDF5 file.}

\item{size}{the maximum length/nchar to use for storage. Default is 32 characters.}
}
\value{
No return value
}
\description{
Create an extensible character HDF5 dataset
}
