% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rna_h5.R
\name{create_ext_h5_float}
\alias{create_ext_h5_float}
\title{Create an extensible float HDF5 dataset}
\usage{
create_ext_h5_float(h5_handle, name)
}
\arguments{
\item{h5_handle}{a connection object created by H5Fopen()}

\item{name}{a character object specifying the full path to the object to modify within the HDF5 file.}
}
\value{
No return value
}
\description{
This will create a 32-bit float dataset compatible with base R numeric floats.
}
