% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rna_h5.R
\name{create_ext_h5_uint}
\alias{create_ext_h5_uint}
\title{Create an extensible unsigned integer HDF5 dataset}
\usage{
create_ext_h5_uint(h5_handle, name, max_val = NULL, bits = 32)
}
\arguments{
\item{h5_handle}{a connection object created by H5Fopen()}

\item{name}{a character object specifying the full path to the object to modify within the HDF5 file.}

\item{max_val}{(optional) the maximum value of the object to write.}

\item{bits}{(optional) the number of bits to use for storing values. Must be 8, 16, 32, or 64.}
}
\value{
No return value
}
\description{
This will generate an unsigned integer dataset within an HDF5 dataset, and will either determine the appropriate bit size
based on max_val or the supplied bits argument. The maximum size of the dataset will be set as H5Sunlimited() (effectively 2^64 - 1).
}
\details{
If both max_val and bits are NULL, will use UINT32 by default.
}
