% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rna_h5.R
\name{expand_h5_dataset}
\alias{expand_h5_dataset}
\title{Expand the size of an HDF5 dataset.}
\usage{
expand_h5_dataset(h5_file, name, new_size)
}
\arguments{
\item{h5_file}{Either a character object specifying the location of an HDF5 file, or a connection created by H5Fopen()}

\item{name}{a character object specifying the full path to the object to modify within the HDF5 file.}

\item{new_size}{a numeric object specifying the new size of the object. Must be greater than the existing size.}
}
\value{
No return.
}
\description{
This will only work up to the size limit set for the
}
