% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_list_path}
\alias{get_list_path}
\title{Retrieve an object from a list using path-style targeting}
\usage{
get_list_path(l, target)
}
\arguments{
\item{l}{a list object}

\item{target}{a character object specifying the "path" to the target.}
}
\value{
a list object
}
\description{
Retrieve an object from a list using path-style targeting
}
\examples{

l <- list(forest = list(country = "USA",
                        maple = list(height = 100)))

maple_height <- get_list_path(l,
                              target = "/forest/maple/height")

forest_list <- get_list_path(l,
                             target = "/forest")

forest_country <- get_list_path(l,
                                target = "/forest/country")

}
