% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{h5_list_add_dgCMatrix}
\alias{h5_list_add_dgCMatrix}
\title{Add a sparse matrix to an h5_list containing 10x genomics data}
\usage{
h5_list_add_dgCMatrix(h5_list, mat, target)
}
\arguments{
\item{h5_list}{a list object, usually generated by running rhdf5::h5dump() on a 10x HDF5 file.}

\item{mat}{a dgCMatrix to add to the h5_list}

\item{target}{a character object specifying the location within the list to write to.}
}
\value{
a list object with the additional matrix added similar to 10x HDF5 matrix format.
}
\description{
Add a sparse matrix to an h5_list containing 10x genomics data
}
