% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{h5_list_cell_metadata}
\alias{h5_list_cell_metadata}
\title{Extract a data.frame of cell metadata from an h5_list object}
\usage{
h5_list_cell_metadata(h5_list)
}
\arguments{
\item{h5_list}{an h5_list object}
}
\value{
a data.frame containing barcodes and all metadata stored in h5_list$matrix$observations.
}
\description{
Extract a data.frame of cell metadata from an h5_list object
}
