% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{h5_list_convert_from_dgCMatrix}
\alias{h5_list_convert_from_dgCMatrix}
\title{Convert the matrix in an h5_list from a sparse matrix back to its original structure}
\usage{
h5_list_convert_from_dgCMatrix(h5_list, target = "matrix")
}
\arguments{
\item{h5_list}{a list object generated by running rhdf5::h5dump() on a 10x HDF5 file, with a matrix converted by h5_list_convert_to_dgCMatrix().}

\item{target}{the group name of the original sparse matrix. Default is "matrix", which is used by 10x.}
}
\value{
a list object
}
\description{
This is very useful for preparing to write back out to an .h5 file.
}
