% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{h5_list_convert_to_dgCMatrix}
\alias{h5_list_convert_to_dgCMatrix}
\title{Convert the matrix in an h5_list from 10x Genomics data to a sparse matrix}
\usage{
h5_list_convert_to_dgCMatrix(h5_list, target = "matrix", feature_names = "id")
}
\arguments{
\item{h5_list}{a list object generated by running rhdf5::h5dump() on a 10x HDF5 file.}

\item{target}{the group name for the sparse matrix. Default is "matrix", which is used by 10x.}

\item{feature_names}{the group name in target/features to use for rownames. Must be either "id" or "name". Default is "id".}
}
\value{
a list object
}
\description{
This is very useful for subsetting the matrix based on column names.
}
