% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{h5_list_transpose}
\alias{h5_list_transpose}
\title{Transpose an h5_list object}
\usage{
h5_list_transpose(h5_list, sparse_matrices = "matrix")
}
\arguments{
\item{h5_list}{an h5_list object generated by running h5dump() on a .h5 file.}

\item{sparse_matrices}{A character vector specifying which objects in the h5_list are sparse matrices which should be transposed.}
}
\value{
a modified h5_list object
}
\description{
This will transpose sparse matrices stored in an h5_list object, and retain the association of observations with columns and features with rows.
}
