% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{h5dims}
\alias{h5dims}
\title{Get dimensions of an object in an HDF5 file}
\usage{
h5dims(h5_file, name)
}
\arguments{
\item{h5_file}{The path to an HDF5 file}

\item{name}{The full name of an object in the HDF5 file.}
}
\value{
a numeric vector containing the dimensions of the target object.
}
\description{
Get dimensions of an object in an HDF5 file
}
