% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{h5exists}
\alias{h5exists}
\title{Check if an object exists in an HDF5 file}
\usage{
h5exists(h5_file, target)
}
\arguments{
\item{h5_file}{A character object specifying the path to a .h5 file}

\item{target}{A character object specifying the name of the object to check (e.g. "/matix/features/id")}
}
\value{
a logical value
}
\description{
Check if an object exists in an HDF5 file
}
