% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{h5ls}
\alias{h5ls}
\title{List objects in an HDF5 file}
\usage{
h5ls(...)
}
\arguments{
\item{...}{parameters passed to rhdf5::h5ls()}
}
\value{
a data.frame listing the contents of the target hdf5 file.
}
\description{
This is a wrapper around rhdf5::h5ls() that adds the full name of each object.
}
