% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_matrix_analysis.R
\name{is_unimodal}
\alias{is_unimodal}
\title{Test if a hash distribution is unimodal}
\usage{
is_unimodal(x, p.cutoff = 0.2)
}
\arguments{
\item{x}{a numeric vector of hash count values}

\item{p.cutoff}{a nuemric value indicating the p-value cutoff for a test of multimodality. If p.value > p.cutoff, the distribution is unimodal.}
}
\value{
a logical value. TRUE if x appears unimodal based on the p-value cutoff.
}
\description{
This function uses the "SI" method provided by multimode::modetest() with B = 20.
}
