% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_matrix_analysis.R
\name{make_singlet_summary}
\alias{make_singlet_summary}
\title{Generate a summary table with the number and fraction of singlets with each HTO barcode}
\usage{
make_singlet_summary(hash_category_table, valid_htos = NULL)
}
\arguments{
\item{hash_category_table}{A hash category table data.frame as produced by categorize_binary_hash_matrix()}

\item{valid_htos}{(optional) a character vector of valid hto sequences. Default is NULL, which will use unique barcodes in hash_category_table$hto_barcode.}
}
\value{
a data.frame with hto_barcode, n_singlets, and frac_singlets.
}
\description{
Generate a summary table with the number and fraction of singlets with each HTO barcode
}
