% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qc_cutoff_barplot}
\alias{qc_cutoff_barplot}
\title{Generate a QC Barplot for a metric at multiple cutoffs}
\usage{
qc_cutoff_barplot(
  meta,
  column_x = "n_umis",
  name_x = "N UMIs",
  cutoffs = c(500, 750, 1000),
  max_y = 30000,
  fill = "purple"
)
}
\arguments{
\item{meta}{A data.frame containing metadata}

\item{column_x}{A character object specifying the metadata to use plotting}

\item{name_x}{A character object specifying a name to display on the x-axis}

\item{cutoffs}{A numeric vector specifying one or more cutoffs to use. Default is c(500, 750, 1000).}

\item{max_y}{A numeric value specifying the maximum value to use for the y-axis. Default is 3e4.}

\item{fill}{A character object specifying the fill color to use for for the violins. Default is "purple".}
}
\value{
a ggplot2 plot object
}
\description{
Generate a QC Barplot for a metric at multiple cutoffs
}
