% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_tables.R
\name{qc_flip}
\alias{qc_flip}
\title{Flip the orientation of a data.frame and return as a list.}
\usage{
qc_flip(df, name_col = NULL)
}
\arguments{
\item{df}{a data.frame or data.frame-like object}

\item{name_col}{(optional) a column in the data.frame used to name each row in the list values}
}
\value{
a nested list object.
}
\description{
Mainly for use in formatting tables for storage in JSON
}
