% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qc_frac_hist_plot}
\alias{qc_frac_hist_plot}
\title{Generate a QC Histogram Plot for a single fractional metric}
\usage{
qc_frac_hist_plot(
  meta,
  column = "n_reads",
  name_x = "N Reads per Cell",
  fill = "dodgerblue",
  target = 0.5,
  y_max = 2000
)
}
\arguments{
\item{meta}{A data.frame containing metadata}

\item{column}{A character object specifying the metadata to display}

\item{name_x}{A character object specifying a name to display on the x-axis}

\item{fill}{A character object specifying the color to use for for the histogram. Default is "dodgerblue".}

\item{target}{A numeric value for a target line to display on the x-axis. Default is 0.5,}

\item{y_max}{A numeric value for the maximum value on the y-axis. Default is 2e3.}
}
\value{
a ggplot2 plot object
}
\description{
Generate a QC Histogram Plot for a single fractional metric
}
