% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qc_scatter_plot}
\alias{qc_scatter_plot}
\title{Generate a QC Scatter Plot for a pair of metrics}
\usage{
qc_scatter_plot(
  meta,
  column_x = "n_reads",
  name_x = "N Reads per Cell",
  column_y = "n_umis",
  name_y = "N UMIs per Cell",
  log_x = TRUE,
  frac_x = FALSE,
  log_y = TRUE,
  frac_y = FALSE,
  show_targets = TRUE,
  color = "dodgerblue"
)
}
\arguments{
\item{meta}{A data.frame containing metadata}

\item{column_x}{A character object specifying the metadata to display on the x-axis}

\item{name_x}{A character object specifying a name to display on the x-axis}

\item{column_y}{A character object specifying the metadata to display on the y-axis}

\item{name_y}{A character object specifying a name to display on the y-axis}

\item{log_x}{A logical indicating whether or not to log10-scale the x-axis. Default is TRUE.}

\item{frac_x}{A logical indicating whether or not to scale the x-axis between 0 and 1. Default is FALSE}

\item{log_y}{A logical indicating whether or not to log10-scale the y-axis. Default is TRUE.}

\item{frac_y}{A logical indicating whether or not to scale the y-axis between 0 and 1. Default is FALSE}

\item{show_targets}{A logical indicating whether or not to plot lines displaying ratios of values. Default is TRUE.}

\item{color}{A character object specifying the color to use for for the points. Default is "dodgerblue".}
}
\value{
a ggplot2 plot object
}
\description{
Generate a QC Scatter Plot for a pair of metrics
}
