% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_tables.R
\name{qc_table}
\alias{qc_table}
\title{Render a nicely-formatted table for QC reports}
\usage{
qc_table(df)
}
\arguments{
\item{df}{A data.frame to format}
}
\value{
an HTML/JavaScript DT object.
}
\description{
Render a nicely-formatted table for QC reports
}
