% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qc_violin_plot}
\alias{qc_violin_plot}
\title{Generate a QC Violin Plot for a metric, grouped by a categorical metadata column.}
\usage{
qc_violin_plot(
  meta,
  category_x = "well_id",
  name_x = "Well ID",
  column_y = "n_reads",
  name_y = "N Reads per Cell",
  log_y = TRUE,
  fill = "skyblue"
)
}
\arguments{
\item{meta}{A data.frame containing metadata}

\item{category_x}{A character object specifying the metadata to use for grouping on the x-axis}

\item{name_x}{A character object specifying a name to display on the x-axis}

\item{column_y}{A character object specifying the metadata to display on the y-axis}

\item{name_y}{A character object specifying a name to display on the y-axis}

\item{log_y}{A logical indicating whether or not to log10-scale the y-axis. Default is TRUE.}

\item{fill}{A character object specifying the fill color to use for for the violins. Default is "skyblue".}
}
\value{
a ggplot2 plot object
}
\description{
Generate a QC Violin Plot for a metric, grouped by a categorical metadata column.
}
