% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_csc_mtx}
\alias{read_csc_mtx}
\title{Read a CITE-seq-Count .mtx directory as a standard R matrix}
\usage{
read_csc_mtx(csc_dir)
}
\arguments{
\item{csc_dir}{a directory containing matrix.mtx, barcodes.tsv, and features.tsv. gzipped versions will also work.}
}
\value{
a matrix
}
\description{
Read a CITE-seq-Count .mtx directory as a standard R matrix
}
