% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{read_h5_cell_meta}
\alias{read_h5_cell_meta}
\title{Read .h5 Cell Metadata}
\usage{
read_h5_cell_meta(h5_file, target = "matrix")
}
\arguments{
\item{h5_file}{the path to an .h5 file in 10x Genomics format}

\item{target}{A matrix object in the .h5 file with a /barcodes object and/or a /target/observations/ sub-group. Default is "matrix".}
}
\value{
A data.frame containing all feature metadata found in /target/barcodes and /target/observations/
}
\description{
Read .h5 Cell Metadata
}
