% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{read_h5_dgCMatrix}
\alias{read_h5_dgCMatrix}
\title{Read the /matrix from a .h5 file as a sparse matrix}
\usage{
read_h5_dgCMatrix(
  h5_file,
  target = "matrix",
  feature_names = "id",
  sample_names = "barcodes",
  index1 = TRUE
)
}
\arguments{
\item{h5_file}{the path to an .h5 file in 10x Genomics format}

\item{target}{a character object specifying the target matrix within the file. Default is "matrix".}

\item{feature_names}{a character object specifying whether to use "id" or "name" for row.names. Default is "id".}

\item{sample_names}{a character object specifying which values to use for col.names. If "barcodes", will use /target/barcodes. Other values will be read from /target/observations/}

\item{index1}{a logical object specifying whether index vectors should start with 0 (FALSE) or 1 (TRUE). Default is TRUE.}
}
\value{
a dgCMatrix of gene expression values.
}
\description{
Read the /matrix from a .h5 file as a sparse matrix
}
