% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{read_h5_feature_meta}
\alias{read_h5_feature_meta}
\title{Read .h5 Feature Metadata}
\usage{
read_h5_feature_meta(h5_file, target = "matrix")
}
\arguments{
\item{h5_file}{the path to an .h5 file in 10x Genomics format}

\item{target}{A matrix object in the .h5 file with a /features/ sub-group. Default is "matrix".}
}
\value{
a data.frame containing all feature metadata found in /target/features/
}
\description{
Read .h5 Feature Metadata
}
