% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{read_h5_sce}
\alias{read_h5_sce}
\title{Read .h5 directly into a SingleCellExperiment object}
\usage{
read_h5_sce(h5_file, target = "matrix", ...)
}
\arguments{
\item{h5_file}{the path to an .h5 file in 10x Genomics format}

\item{target}{A matrix object in the .h5 file with a /features/ sub-group. Default is "matrix".}

\item{...}{Additional parameters passed to \code{\link{SingleCellExperiment::SingleCellExperiment()}}}
}
\value{
a SingleCellExperiment Class object
}
\description{
By default, matrix data will be stored in the "counts" assay.
}
\details{
CITE-seq data will be automatically detected and stored in the altExp called "ADT"
}
