% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_tenx_metrics}
\alias{read_tenx_metrics}
\title{Read and correct formatting of a 10x metrics_summary.csv file}
\usage{
read_tenx_metrics(metrics_csv)
}
\arguments{
\item{metrics_csv}{path to a metrics_summary.csv file generated by cellranger}
}
\value{
a data.frame of cellranger run metrics
}
\description{
Read and correct formatting of a 10x metrics_summary.csv file
}
