% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{reduce_h5_list}
\alias{reduce_h5_list}
\title{Reduce a list-of-lists of h5_list objects to a single, concatenated object.}
\usage{
reduce_h5_list(h5_ll, sparse_matrices = c("matrix"))
}
\arguments{
\item{h5_ll}{A list-of-lists containing h5_list objects generated by rhdf5::h5dump() or similar.}

\item{sparse_matrices}{A vector of object names within each list that should be treated as dgCMatrices for combining. Default is "matrix", which is used by 10x.}
}
\value{
An h5_list object with the structure of a single h5_list within the h5_ll object.
}
\description{
Reduce a list-of-lists of h5_list objects to a single, concatenated object.
}
