% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_matrix_analysis.R
\name{select_hash_cutoff}
\alias{select_hash_cutoff}
\title{Select a cutoff for hash counts using automatic, k-means based splitting}
\usage{
select_hash_cutoff(
  x,
  use_median_cut = FALSE,
  min_cut = 10,
  min_fc = 2,
  seed = 3030
)
}
\arguments{
\item{x}{a numeric vector of hash count values}

\item{use_median_cut}{a logical value indicating whether or not to use the median value of x to set a minimum threshold. Default is FALSE.}

\item{min_cut}{a numeric value for the minimum number of counts to consider. Default is 10.}

\item{min_fc}{a numeric value for the minimum expected fold change between the centers of the two clusters (positive and negative). Default is 2.}

\item{seed}{a value to use as a random seed for k-means clustering. Default is 3030.}
}
\value{
a numeric value to use as a cutoff. If no values are above min_cut, returns max(x), otherwise returns max(x) for non-passing values.
}
\description{
Select a cutoff for hash counts using automatic, k-means based splitting
}
