% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_list_path}
\alias{set_list_path}
\title{Set a list value using path-style targeting}
\usage{
set_list_path(l, target, value)
}
\arguments{
\item{l}{a list object}

\item{target}{a character object specifying the "path" to the target.}

\item{value}{an object or value to insert at the target location}
}
\value{
a list object
}
\description{
Set a list value using path-style targeting
}
\examples{

l <- list(forest = list(country = "USA",
                        maple = list(height = 100)))

l <- set_list_path(l,
                   target = "/forest/spruce/height",
                   value = 30)

l <- set_list_path(l,
                   target = "/forest/maple/diameter",
                   value = 6)

l <- set_list_path(l,
                   target = "/valley/country",
                   value = "Canada")

}
