% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{split_h5_list_by_hash}
\alias{split_h5_list_by_hash}
\title{Split a 10x HDF5 file based on HTOparser results}
\usage{
split_h5_list_by_hash(h5_list)
}
\arguments{
\item{h5_list}{a list object generated by h5dump(), which has hash labels obtained from add_h5_list_hash_results().}
}
\value{
A list object with contents of the input h5_file separated by each hto_barcode. Also includes multiplets.
}
\description{
Split a 10x HDF5 file based on HTOparser results
}
