% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{stm}
\alias{stm}
\title{Write a stderr message with a leading date/time stamp}
\usage{
stm(x)
}
\arguments{
\item{x}{a character object with the message to display}
}
\value{
no return
}
\description{
Write a stderr message with a leading date/time stamp
}
