% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{strip_1d_array_recursive}
\alias{strip_1d_array_recursive}
\title{Convert all 1D Arrays in a list object to vectors recursively}
\usage{
strip_1d_array_recursive(x)
}
\arguments{
\item{x}{a list object}
}
\value{
a list object with all 1D arrays converted.
}
\description{
Convert all 1D Arrays in a list object to vectors recursively
}
