% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{subset_h5_list_by_observations}
\alias{subset_h5_list_by_observations}
\title{Subset a h5_list object based on a set of observations.}
\usage{
subset_h5_list_by_observations(
  h5_list,
  match_values,
  match_target = "barcodes",
  sparse_matrices = "matrix"
)
}
\arguments{
\item{h5_list}{a list object generated by running rhdf5::h5dump() on a 10x HDF5 file.}

\item{match_values}{a vector of values to use for filtering}

\item{match_target}{a character object specifying which values to use for filtering in the h5_list. If "barcodes", uses matrix/barcodes; Otherwise, looks in matrix/observations for matching values}

\item{sparse_matrices}{a character vector specifying which list objects to treat as sparse matrices. If these aren't currently dgCMatrices, they will be converted by h5_list_convert_to_dgCMatrix for filtering.}
}
\value{
a list object
}
\description{
Subset a h5_list object based on a set of observations.
}
