% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalseq_references.R
\name{totalseq_a_biotin}
\alias{totalseq_a_biotin}
\title{Load BioLegend TotalSeq A reference table for biotin hash tag oligos (HTO)}
\usage{
totalseq_a_biotin()
}
\value{
a data.table with reference information
}
\description{
Load BioLegend TotalSeq A reference table for biotin hash tag oligos (HTO)
}
