% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalseq_references.R
\name{totalseq_a_human}
\alias{totalseq_a_human}
\title{Load BioLegend TotalSeq A reference table for human hash tag oligos (HTO)}
\usage{
totalseq_a_human()
}
\value{
a data.table with reference information
}
\description{
Note: there is no Hashtag 11 in the human HTO set.
}
