% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rna_h5.R
\name{write_10x_h5_container}
\alias{write_10x_h5_container}
\title{Generate an empty HDF5 container to store RNA-seq count data in 10x format}
\usage{
write_10x_h5_container(
  h5_file,
  cell_barcode_length = 15,
  data_type = "integer",
  data_bits = 16,
  indices_bits = 16,
  indptr_bits = 32
)
}
\arguments{
\item{cell_barcode_length}{18}

\item{data_type}{integer}

\item{data_bits}{16}

\item{indices_bits}{16}

\item{indptr_bits}{32}
}
\value{

}
\description{
Generate an empty HDF5 container to store RNA-seq count data in 10x format
}
